+++ apps/ymir-sdl3/src/app/settings.cpp
@@ -706,6 +706,7 @@ void Settings::ResetToDefaults() {
     video.autoResizeWindow = false;
     video.displayVideoOutputInWindow = false;
     video.syncInWindowedMode = true;
+    video.reduceLatency = true;
     video.fullScreen = false;
     video.doubleClickToFullScreen = false;
     video.useFullRefreshRateInFullScreen = false;
@@ -1055,6 +1056,7 @@ SettingsLoadResult Settings::Load(const std::filesystem::path &path) {
         Parse(tblVideo, "AutoResizeWindow", video.autoResizeWindow);
         Parse(tblVideo, "DisplayVideoOutputInWindow", video.displayVideoOutputInWindow);
         Parse(tblVideo, "SyncInWindowedMode", video.syncInWindowedMode);
+        Parse(tblVideo, "ReduceLatency", video.reduceLatency);
         Parse(tblVideo, "FullScreen", video.fullScreen);
         Parse(tblVideo, "DoubleClickToFullScreen", video.doubleClickToFullScreen);
         Parse(tblVideo, "UseFullRefreshRateInFullScreen", video.useFullRefreshRateInFullScreen);
@@ -1466,6 +1468,7 @@ SettingsSaveResult Settings::Save() {
             {"AutoResizeWindow", video.autoResizeWindow},
             {"DisplayVideoOutputInWindow", video.displayVideoOutputInWindow},
             {"SyncInWindowedMode", video.syncInWindowedMode},
+            {"ReduceLatency", video.reduceLatency},
             {"FullScreen", video.fullScreen.Get()},
             {"DoubleClickToFullScreen", video.doubleClickToFullScreen},
             {"UseFullRefreshRateInFullScreen", video.useFullRefreshRateInFullScreen},